Update phpunit and remove support for PHP 5.6 and 7.0
PHPUnit 6 introduced namespaces, so we cannot maintain a single testset that works both with PHPUnit 5 and PHPUnit 6+. Since both PHPUnit 5 and PHP 5 have now reached end-of-life, it's time to move on.

The sru-client library might still work on older PHP versions, but official support is removed since it cannot be automatically tested anymore.